fix: fixed news, layout, styles
