+++ package.json
@@ -12,6 +12,7 @@
 		"react-dom": "^18.2.0",
 		"react-scripts": "5.0.1",
 		"sass": "^1.69.1",
+		"swiper": "^10.3.1",
 		"web-vitals": "^2.1.4"
 	},
 	"scripts": {
