--- src/components/Main/AboutProject/AboutProject.css
@@ -1,21 +1,0 @@
-.about-project {
-	margin: 100px 130px 0;
-	width: 1660px;
-}
-
-.about-project__title {
-	margin: 0;
-	font-size: 56px;
-	font-style: normal;
-	font-weight: 500;
-	line-height: 100%;
-}
-
-.about-project__description {
-	font-size: 34px;
-	font-style: normal;
-	font-weight: 400;
-	line-height: 40px;
-	margin: 0;
-	padding-top: 50px;
-}
