--- src/components/Main/AboutProject/AboutProject.js
@@ -1,16 +1,0 @@
-import './AboutProject.css';
-
-function AboutProject() {
-	return (
-		<section className="about-project">
-			<h2 className="about-project__title">О нас</h2>
-			<p className="about-project__description">
-				Платформа создана, чтобы объединить усилия волонтёров, городских
-				администраций <br /> и организаторов мероприятий и внести совместный
-				вклад в развитие городской среды.
-			</p>
-		</section>
-	);
-}
-
-export default AboutProject;
