+++ src/components/Main/AboutProject/AboutProject.scss
@@ -1,6 +1,29 @@
-.our-values__container {
-	width: 1660px;
+.about-project {
 	margin: 100px 130px 0;
+	width: 1660px;
+}
+
+.about-project__title {
+	margin: 0;
+	font-size: 56px;
+	font-style: normal;
+	font-weight: 500;
+	line-height: 100%;
+}
+
+.about-project__description {
+	font-size: 34px;
+	font-style: normal;
+	font-weight: 400;
+	line-height: 40px;
+	margin: 0;
+	padding-top: 50px;
+}
+
+.about-project__container {
+	/* width: 1660px;
+	margin: 100px 130px 0; */
+	margin-top: 100px;
 	padding: 0;
 }
 .our-values__cards {
