+++ src/components/Main/JoinButtons/JoinButtons.scss
@@ -49,3 +49,7 @@
 	right: 40px;
 	bottom: 40px;
 }
+
+.join__button:hover {
+	opacity: 0.7;
+}
