--- src/components/Main/OurValues/OurValues.stories.js
@@ -1,8 +1,0 @@
-import OurValues from './OurValues';
-
-export default {
-	title: 'OurValues',
-	component: OurValues,
-};
-
-export const Default = (arg) => <OurValues {...arg} />;
