+++ src/components/Main/Promo/Promo.scss
